+++ src/tm_devices/drivers/source_measure_units/source_measure_unit.py
@@ -7,12 +7,19 @@ from abc import ABC
 
 from tm_devices.drivers.device import Device
 from tm_devices.helpers import DeviceTypes
+from tm_devices.helpers import ReadOnlyCachedProperty as cached_property  # noqa: N813
 
 
 class SourceMeasureUnit(Device, ABC):
     """Base Source Measure Unit (SMU) device driver."""
 
-    _DEVICE_TYPE = DeviceTypes.SMU.value
+    ################################################################################################
+    # Properties
+    ################################################################################################
+    @cached_property
+    def device_type(self) -> str:
+        """Return a string representing the device type."""
+        return DeviceTypes.SMU.value
 
     ################################################################################################
     # Public Methods