+++ tests/test_unsupported_device_type.py
@@ -15,7 +15,9 @@ from tm_devices.helpers import ReadOnlyCachedProperty as cached_property  # noqa
 class CustomUnsupportedDeviceUnitTestOnly(PIControl, Device):
     """A custom device that is not one of the officially supported devices for unit tests."""
 
-    _DEVICE_TYPE = "CustomDeviceType"
+    @cached_property
+    def device_type(self) -> str:  # noqa: D102
+        return "CustomDeviceType"
 
     def _get_errors(self) -> Tuple[int, Tuple[str, ...]]:
         return 0, ()